pyproject.toml
[project]
name = "localaicoder"
version = "2.6.0"
description = "This is an Ai assistant pro programmers and coders."
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.5.1",
    "python-dotenv>=1.1.1",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "huggingface-hub>=0.20.0",
    "accelerate>=0.20.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
]
